Split test_migration.py into focused test files
